pyproject.toml
[tool.poetry]
name = "papers-please"
version = "0.1.0"
description = ""
authors = ["Maksim Vasilev <maxim.vasilev@jetbrains.com>"]
packages = [{ include = "papers_please" }]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.1.3"
python-anticaptcha = "^0.7.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
